requirements.txt
dash
plotly.express
pandas